+++ controls.h
@@ -4,8 +4,8 @@
 
 extern int show_center_mark; //for renderer to show center mark
 extern int show_canvas_rotation_point; //for renderer to show another reference mark
-extern float canvas_rotation_point_x;
-extern float canvas_rotation_point_y;
+extern double canvas_rotation_point_x;
+extern double canvas_rotation_point_y;
 extern int global_testA;
 extern int global_testB;
 extern int global_testC;
