+++ images.h
@@ -14,12 +14,12 @@ typedef struct {
   SDL_Texture *texture;
   int width;
   int height;
-  float x;
-  float y;
-  float r;  // rotation angle, and rotation points
-  float rx;
-  float ry;
-  float z;
+  double x;
+  double y;
+  double r;  // rotation angle, and rotation points
+  double rx;
+  double ry;
+  double z;
   uint8_t opacity;
   int crop_top;
   int crop_right;
@@ -32,32 +32,32 @@ typedef struct {
 } Image;
 
 typedef struct {
-  float aX;
-  float aY;
-  float bX;
-  float bY;
-  float cX;
-  float cY;
-  float dX;
-  float dY;
+  double aX;
+  double aY;
+  double bX;
+  double bY;
+  double cX;
+  double cY;
+  double dX;
+  double dY;
 } rectangleCorners;
 
 extern Image *images;
 
-extern float imrefAx;
-extern float imrefAy;
-extern float imrefBx;
-extern float imrefBy;
-extern float imrefCx;
-extern float imrefCy;
-extern float imrefDx;
-extern float imrefDy;
+extern double imrefAx;
+extern double imrefAy;
+extern double imrefBx;
+extern double imrefBy;
+extern double imrefCx;
+extern double imrefCy;
+extern double imrefDx;
+extern double imrefDy;
 
 void images_free(); //to be called when program exits
 void images_load_dir(const char *directory);  // load all images from directory
 void images_arrange_in_grid();
 
-int image_point_on(float x, float y); //returns images[] index of image currently under the given point, or -1 if none
+int image_point_on(double x, double y); //returns images[] index of image currently under the given point, or -1 if none
 
 void images_render();
 
@@ -66,7 +66,7 @@ rectangleCorners image_find_corners(int si);
 void image_to_on_bottom(int imi);
 void image_to_on_top(int imi);
 void image_drag_screen_by(int imi, int dx, int dy);
-void image_zoom_by(int imi, float zoom_factor);
+void image_zoom_by(int imi, double zoom_factor);
 void image_zoom_reset(int imi);
-void image_rotate_by(int imi, float dr);
-void image_rotate_snap(int imi, float sr);
+void image_rotate_by(int imi, double dr);
+void image_rotate_snap(int imi, double sr);
